Sprint 332 (0.50.95): documentation compaction + structural audit
- DOC-332-01: Delete 4 stale docs, create ARCHIVE.md (18k lines),
  compact backlog/checklist/gap_audit (18k->~400 lines),
  update IMPLEMENTATION_SUMMARY.md to v0.50.94
- STR-332-02: 3 violations partitioned (709->dir, 670->dir, 536->dir);
  ZERO files > 500 lines workspace-wide
- Update CHANGELOG.md, OPTIMIZATION.md, README.md with Sprint 331-332
- Verify: clippy=0 warnings, ritk-core=1408 pass, ritk-registration=547 pass

+++ crates/ritk-core/src/filter/surface/tests_marching_cubes.rs
@@ -1,5 +1,6 @@
 //! Tests for marching_cubes
 //! Extracted to keep the 500-line structural limit.
+#![allow(clippy::identity_op, clippy::erasing_op)]
 use super::mc_tables::EDGE_TABLE;
 use super::*;
 use gaia::domain::core::index::VertexId;
@@ -178,9 +179,9 @@ fn origin_shifts_vertex_positions() {
     let verts_orig = collect_vertices(&mesh_orig);
     let verts_shift = collect_vertices(&mesh_shift);
     for (u, s) in verts_orig.iter().zip(verts_shift.iter()) {
-        assert!((s[0] - u[0] - origin[0] as f64).abs() < 1e-4);
-        assert!((s[1] - u[1] - origin[1] as f64).abs() < 1e-4);
-        assert!((s[2] - u[2] - origin[2] as f64).abs() < 1e-4);
+        assert!((s[0] - u[0] - origin[0]).abs() < 1e-4);
+        assert!((s[1] - u[1] - origin[1]).abs() < 1e-4);
+        assert!((s[2] - u[2] - origin[2]).abs() < 1e-4);
     }
 }
